// src/app/core/models/pioms/agentAttendance.ts
export class AgentAttendance {
    agentAttendanceID: number;
    complainID: number;
    complainNo: string;
    agentID: number | null;
    accusedBPNumber: string;
    dutyPlace: string;
    firstAttendanceTime: string | null;
    secondAttendanceTime: string | null;
    firstLatitude: number | null;
    firstLongitude: number | null;
    secondLatitude: number | null;
    secondLongitude: number | null;
    remarks: string;
    fromDate: Date;
    toDate: Date;
     bpNumber: string;

   
}